requirements.txt
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.1
pandas==2.2.2
python-dotenv==1.0.1
gradio==4.44.0